requirements.txt
absl-py==0.7.1
astor==0.8.0
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.16.4
pkg-resources==0.0.0
protobuf==3.8.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tweet-preprocessor==0.5.0
Werkzeug==0.15.4
wrapt==1.11.2
